Introduce Python alike package namespace, turn run.py into an entrypoint
