+++ roop/analyser.py
@@ -1,5 +1,5 @@
 import insightface
-import core.globals
+import roop.globals
 
 FACE_ANALYSER = None
 
@@ -7,7 +7,7 @@ FACE_ANALYSER = None
 def get_face_analyser():
     global FACE_ANALYSER
     if FACE_ANALYSER is None:
-        FACE_ANALYSER = insightface.app.FaceAnalysis(name='buffalo_l', providers=core.globals.providers)
+        FACE_ANALYSER = insightface.app.FaceAnalysis(name='buffalo_l', providers=roop.globals.providers)
         FACE_ANALYSER.prepare(ctx_id=0, det_size=(640, 640))
     return FACE_ANALYSER
