+++ roop/globals.py
